+++ components/Hooks/useClickAway.ts
@@ -15,7 +15,6 @@
  * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 import React, {useEffect} from "react"
 
 const useClickAway = (container: React.RefObject<HTMLDivElement>, enabled: boolean, onClickAway: () => void) => {