+++ components/PromptBar/PromptBar.context.tsx
@@ -15,7 +15,6 @@
  * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 import {Dispatch, createContext} from "react"
 
 import {PromptBarInitialState} from "./PromptBar.state"