+++ components/Sidebar/Sidebar.tsx
@@ -15,7 +15,6 @@
  * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 import {IconFolderPlus, IconMistOff, IconTextPlus} from "@tabler/icons-react"
 import {ReactNode} from "react"
 import {useTranslation} from "react-i18next"