+++ pages/api/home/home.state.tsx
@@ -15,7 +15,6 @@
  * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 import {Conversation, Message} from "@/types/chat"
 import {ErrorMessage} from "@/types/error"
 import {FolderInterface} from "@/types/folder"