+++ utils/app/conversations.ts
@@ -15,18 +15,27 @@
  * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 import {v4 as uuidv4} from "uuid"
 
 import {Conversation} from "@/types/chat"
 import {maxOutputTokensForModel} from "@/types/openai"
-import {OPENAI_API_MAX_TOKENS, OPENAI_DEFAULT_REASONING_EFFORT, OPENAI_DEFAULT_SYSTEM_PROMPT, OPENAI_DEFAULT_TEMPERATURE} from "@/utils/app/const"
+import {
+  OPENAI_API_MAX_TOKENS,
+  OPENAI_DEFAULT_REASONING_EFFORT,
+  OPENAI_DEFAULT_SYSTEM_PROMPT,
+  OPENAI_DEFAULT_TEMPERATURE
+} from "@/utils/app/const"
 import {localStorageSafeGetItem, localStorageSafeRemoveItem, localStorageSafeSetItem} from "@/utils/app/storage"
 
 export const LOCAL_STORAGE_SELECTED_CONVERSATION = "selectedConversation"
 export const STORAGE_KEY_HISTORY = "history"
 
-export const createNewConversation = (name: string, modelId: string, temperature: number, reasoningEffort: string): Conversation => {
+export const createNewConversation = (
+  name: string,
+  modelId: string,
+  temperature: number,
+  reasoningEffort: string
+): Conversation => {
   return {
     id: uuidv4(),
     name: name,